FIX: constant log in BTD, incorrect use of serr in DistanceGrid, misc warnings

+++ modules/SofaGeneralTopology/BezierTetrahedronSetTopologyContainer.cpp
@@ -584,7 +584,7 @@ bool BezierTetrahedronSetTopologyContainer::checkBezierPointTopology()
 	size_t nTetras,elem;
 	BezierDegreeType degree=d_degree.getValue();
 	// check the total number of vertices.
-    assert((size_t)getNbPoints()==(getNumberOfTetrahedralPoints()+getNumberOfEdges()*(degree-1)+getNumberOfTriangles()*(degree-1)*(degree-2)/2+getNumberOfTetrahedra()*(degree-1)*(degree-2)*(degree-3)/6));
+	assert((int)getNbPoints()==(int)(getNumberOfTetrahedralPoints()+getNumberOfEdges()*(degree-1)+getNumberOfTriangles()*(degree-1)*(degree-2)/2+getNumberOfTetrahedra()*(degree-1)*(degree-2)*(degree-3)/6));
 	sofa::helper::vector<TetrahedronBezierIndex> tbiArray=getTetrahedronBezierIndexArray();
 	VecPointID indexArray;
 	BezierTetrahedronPointLocation location; 
@@ -593,7 +593,7 @@ bool BezierTetrahedronSetTopologyContainer::checkBezierPointTopology()
 		indexArray.clear();
 		getGlobalIndexArrayOfBezierPointsInTetrahedron(nTetras,indexArray);
 		// check the number of control points per tetrahedron is correct
-        assert(indexArray.size()==(size_t)((4+6*(degree-1)+2*(degree-1)*(degree-2)+(degree-1)*(degree-2)*(degree-3)/6)));
+		assert((int)indexArray.size()==(int)(4+6*(degree-1)+2*(degree-1)*(degree-2)+(degree-1)*(degree-2)*(degree-3)/6));
 		for(elem=0;elem<indexArray.size();++elem) {
 			size_t globalIndex=getGlobalIndexOfBezierPoint(nTetras,tbiArray[elem]);
 			// check that getGlobalIndexOfBezierPoint and getGlobalIndexArrayOfBezierPointsInTetrahedron give the same answer
